+++ src/lib/extra_index.js
@@ -7,6 +7,7 @@ import dagre from 'cytoscape-dagre';
 import euler from 'cytoscape-euler';
 import klay from 'cytoscape-klay';
 import spread from 'cytoscape-spread';
+import svg from 'cytoscape-svg';
 
 CytoscapeJS.use(coseBilkent);
 CytoscapeJS.use(cola);
@@ -14,6 +15,7 @@ CytoscapeJS.use(dagre);
 CytoscapeJS.use(euler);
 CytoscapeJS.use(klay);
 CytoscapeJS.use(spread);
+CytoscapeJS.use(svg);
 
 export {
     Cytoscape